async-load workshop enrollment enrichment
